Reverse the ordering of Discord and Gitter badges.
I joined Gitter yesterday and was directed to Discord which has more activity. I see others made the same mistakes earlier that day. I noticed later it was explained in the README.md, but since I only read the badges left-to-right while looking for support I didn't find it initially. My guess is this will reduce the number of people who land in Gitter and need to be redirected.